DevOps Development/Cloud Cost report/requirements.txt
functions-framework==3.5.0
google-cloud-billing==1.13.0
google-cloud-bigquery==3.14.0
google-cloud-secret-manager==2.16.4
requests==2.31.0
